+++ GLImp/Model Importing/Model/Vertex.cs
@@ -9,6 +9,6 @@ namespace Mila.Data
     /// </summary>
     public class Vertex
     {
-        public float x = 0.0f, y = 0.0f, z = 0.0f;
+        public double x = 0.0f, y = 0.0f, z = 0.0f;
     }
 }
